Finish the module rename
